Website now one page with static HTML and CSS
This website has gone through a major change to make it a lot more professional and easier for recruiters to use to look through and determine whether or not I am eligible for some jobs I could apply to. In addition, it makes it much easier for me to manage my projects and publish them as needed.